V2.0.0
CPU1 Bootloader & Application

+++ source/inc/bsp/bsp_filesystem.h
@@ -11,6 +11,6 @@
 void test_fatfs(void);
 FRESULT ChangeToDirectory(char *pcDirectory, char *now_dir, unsigned long *pulReason);
 void init_fileSystem_buffer(void);
-int16 read_file(Uint16 *address, Uint16 *buffer, Uint16 read_size, Uint16 index);
+int16 read_file(Uint16 *address, Uint16 *buffer, Uint16 read_size, Uint32 index);
 
 #endif /* INC_BSP_BSP_FILESYSTEM_H_ */
